+++ src/main/web_src/src/components/fagsystem/inntektstub/form/partials/inntektsinformasjonForm.tsx
@@ -1,13 +1,11 @@
 import React, { useState } from 'react'
 import { FormikProps } from 'formik'
-import { FormikTextInput, TextInput } from '~/components/ui/form/inputs/textInput/TextInput'
+import { FormikTextInput } from '~/components/ui/form/inputs/textInput/TextInput'
 import { InntektstubVirksomhetToggle } from './inntektstubVirksomhetToggle'
 import InntektsinformasjonLister from './inntektsinformasjonLister/inntektsinformasjonLister'
 import InntektsendringForm from './inntektsendringForm'
-import ReactDatepicker from 'react-datepicker'
-import { Label } from '~/components/ui/form/inputs/label/Label'
-import { InputWrapper } from '~/components/ui/form/inputWrapper/InputWrapper'
 import _get from 'lodash/get'
+import { Monthpicker } from '~/components/ui/form/inputs/monthpicker/Monthpicker'
 
 interface InntektsinformasjonForm {
 	path: string
@@ -29,39 +27,16 @@ export default ({ path, formikBag }: InntektsinformasjonForm) => {
 		)
 	}
 
-	const getFeilmelding = (formikProps: FormikProps<{}>, formikPath: string) => {
-		const feilmelding = _get(formikProps.errors, formikPath)
-		return feilmelding ? { feilmelding: feilmelding } : null
-	}
-
 	return (
 		<div key={path}>
 			<div className="flexbox--flex-wrap">
-				<InputWrapper>
-					<Label
-						name={`${path}.sisteAarMaaned`}
-						label={'År/måned'}
-						feil={getFeilmelding(formikBag, `${path}.sisteAarMaaned`)}
-					>
-						<ReactDatepicker
-							className={'skjemaelement__input'}
-							locale="nb"
-							dateFormat="yyyy-MM"
-							selected={date}
-							onChange={handleDateChange}
-							placeholderText={'yyyy-MM'}
-							showMonthYearPicker
-							customInput={
-								<TextInput
-									icon="calendar"
-									feil={getFeilmelding(formikBag, `${path}.sisteAarMaaned`)}
-								/>
-							}
-							dropdownMode="select"
-							autoComplete="off"
-						/>
-					</Label>
-				</InputWrapper>
+				<Monthpicker
+					formikBag={formikBag}
+					name={`${path}.sisteAarMaaned`}
+					label="År/måned"
+					date={date}
+					handleDateChange={handleDateChange}
+				/>
 				<FormikTextInput
 					name={`${path}.antallMaaneder`}
 					label="Generer antall måneder"
